requirements.txt
boto3==1.4.3
botocore==1.4.90
cffi==1.9.1
cryptography==1.7.1
Django==1.10.4
django-common-helpers==0.9.0
django-cron==0.5.0
django-crontab==0.7.1
django-geoposition==0.3.0
django-jquery==3.1.0
django-pdb==0.5.1
docutils==0.13.1
elasticsearch==2.4.0
enum34==1.1.6
futures==3.0.5
geopy==1.11.0
idna==2.2
ipaddress==1.0.17
jmespath==0.9.0
pyasn1==0.1.9
pycparser==2.17
pyOpenSSL==16.2.0
python-dateutil==2.6.0
requests==2.9.1
s3transfer==0.1.10
six==1.10.0
urllib3==1.19.1
